Update vignette
Deprecate nu function.

Add "type" option to coef and sdev to return result as flat vector or named list.

Add "type" option to predict so that link estimates can be requested instead of response.

Add warning to mle procedures, that "fnscale" will be disregarded if it is set.

Remove leverage.zicmp since it was never implemented or worked out.