Update vignette
Warn if qcmp and qzicmp arguments are too close to 1.

Update warnings if y is truncated too soon. Only return the attained
diff if we reach the "second" stage of the loop, where a geometric
series can be used to bound the original series.

Fix details on Geometric special cases.